test_allocation_failure: Tighten memory to increase test pass chances
In order to create an out of memory condition in the right time, the
test begins with a loop that exhaust all `fragmented_temporary_buffer::default_fragment_size`
chunks of memory. Just after the loop exhaust all memory it releases one chunk out of
the implicit assumption that it will be used for continuation creation. However, this assumption
is not necessarily true as the memory exhaustion is not tight enough.
1. At the end of the loop we can only be sure that Available_Memory <
`fragmented_temporary_buffer::default_fragment_size`
2. After erasing the last element we can actually end up with
Available_Memory >
`fragmented_temporary_buffer::default_fragment_size`*2
3. The continuation seams (empirically) to be able to use the <
`fragmented_temporary_buffer::default_fragment_size`
memory for allocation
which can have us end up with a free 128K buffer and the additional last
fragment after the continuation allocation.
In order to tighten the memory shortage we changed the following:
Right after the `fragmented_temporary_buffer::default_fragment_size` allocation loop breaks,
save an iterator to the last `fragmented_temporary_buffer::default_fragment_size` chunk,
assume that there are (`fragmented_temporary_buffer::default_fragment_size` - 1) available bytes in the system and try
to allocate them all. Release one chunk by resetting the pointer instead of erasing the
list element (which might return some more extra bytes to the heap).

Tested with prints that confirmed that indeed there is more memory to be
allocated. Except for proving that there is more memory to be allocated,
I didn't prove that it can be used for continuation allocation, this was
deduced from the previous code flow and path that generated the unexpected
success exception.

Fixes #9391

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>