loading_cache: account new generation partition evictions
Provide a template parameter to provide a static callbacks object to
increment a counter of evictions from a new generation partition.

Entries being evicted from the cache while still in a new generation
partition indicates a not efficient usage of the cache and should be
investigated.

This patch instruments authorized_prepared_statements_cache and a
prepared_statements_cache objects to provide non-empty callbacks.

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>