murmur3: fix inconsistent token for empty partition key
Traditionally in Scylla and in Cassandra, an empty partition key is mapped
to minimum_token() instead of the empty key's usual hash function (0).
The reasons for this are unknown (to me), but one possibility is that
having one known key that maps to the minimal token is useful for
various iterations.

In murmur3_partitioner.cc we have two variants of the token calculation
function - the first is get_token(bytes_view) and the second is
get_token(schema, partition_key_view). The first includes that empty-
key special case, but the second was missing this special case!

As Kamil first noted in #9352, the second variant is used when looking
up partitions in the index file - so if a partition with an empty-string
key is saved under one token, it will be looked up under a different
token and not found. I reproduced exactly this problem when fixing
issues #9364 and #9375 (empty-string keys in materialized views and
indexes) - where a partition with an empty key was visible in a
full-table scan but couldn't be found by looking up its key because of
the wrong index lookup.

I also tried an alternative fix - changing both implementations to *not*
special-case the empty key (and return token 0 for it) - but this version
didn't fix the lookup bug. To be honest, I'm not sure why... But anyway,
given the fact that I found multiple places in the Scylla code that assume
that dht::minimum_token() and partition_key::make_empty() are paired,
I think it's safest to leave this exception in place.

This patch does not risk a backward-incompatible disk format changes
for two reasons:
1. The fixed code apparently is only used for lookup, not writing.
2. In any case, in the current Scylla, there was no valid case where an
   empty partition key may appear. CQL and Thrift forbid such keys, and
   materialized-views and indexes also (incorrectly - see #9364, #9375)
   drop such rows.

When #9364 and #9375 will be fixed by the next patch, empty partition keys
will start appearing in Scylla (in materialized views and in the
materialized view backing a secondary index), and this fix will become
important.

Fixes #9352
Refs #9364
Refs #9375

Signed-off-by: Nadav Har'El <nyh@scylladb.com>