database: add _registry member
We want to make the schema registry a member of database. To help
prepare the code to that in an incremental manner, we add a registry
field to database, referencing the global registry for now. Once all the
code is migrated to to obtain the registry from the database instead of
using the global instance, we can retire the latter.