schema_registry: make it a peering sharded service
Most of the codebase has been migrated to obtain the registry from  the
database. The last major user -- global_schema_ptr -- however is
trickier to migrate. This class needs to access the registry's
counterpart on a remote shard. To enable this use-case we make the
schema registry peering sharded service, declared and initialized in
main and passed to the database as a reference (to the shard-local
instance).
For now we keep the thread-local variable too, initialized to the
sharded service's local instance in main. This is to facilitate
migrating the global_schema_ptr in a separate patch.